pyproject.toml
[tool.poetry]
name = "gnome-next-meeting-applet"
version = "2.1.1"
description = "Show your next meeetings in an applet"
authors = ["Chmouel Boudjnah <chmouel@chmouel.com>"]
maintainers = ["Chmouel Boudjnah <chmouel@chmouel.com>"]
license = "MIT"
readme = "README.md"
homepage = ""
keywords = [
    "calendar", "Gnome",
]
packages = [
    { include = "gnma" }
]

include = [
 "data/"
]

[tool.poetry.dependencies]
python = "^3.8"
PyGObject = "^3.0.0"
python-dateutil = "*"
PyYAML = "*"
humanize = "*"
dbus-python = "^1.2.18"

[tool.poetry.dev-dependencies]
pytest = "*"
yapf = "*"
toml = "*"
pylint = "*"

[tool.poetry.scripts]
gnome-next-meeting-applet = "gnma.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
issues = "https://github.com/chmouel/gnome-next-meeting-applet/issues"
